% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepanel.mountainplot}
\alias{prepanel.mountainplot}
\title{The prepanel function for mountainplot}
\usage{
prepanel.mountainplot(x, ...)
}
\arguments{
\item{x}{The data to be plotted.}

\item{...}{Other arguments}
}
\description{
The prepanel function for mountainplot
}

