% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountain.R
\name{panel.mountainplot}
\alias{panel.mountainplot}
\title{The panel function for mountainplot}
\usage{
panel.mountainplot(x, type = "s", groups = NULL, ref = TRUE, ...)
}
\arguments{
\item{x}{The data to be plotted.}

\item{type}{The type of ecdf line to use.  Default is 's' square.}

\item{groups}{Variable to use for grouping}

\item{ref}{If TRUE, draw horizontal reference lines at 0,1}

\item{...}{Other arguments}
}
\description{
The panel function for mountainplot
}

